ECS refactor
