+++ ecs/src/plex/ecs/entity_registry.cpp
@@ -1,0 +1,17 @@
+#include "plex/ecs/entity_registry.h"
+
+namespace plex
+{
+EntityRegistry::EntityRegistry()
+{
+  storages_.resize(MaxArchetypes);
+}
+
+EntityRegistry::~EntityRegistry()
+{
+  for (auto storage : storages_)
+  {
+    if (storage) delete storage;
+  }
+}
+} // namespace plex
