+++ ecs/test/unit/ecs/view_relations_tests.cpp
@@ -1,12 +1,92 @@
-#include "plex/ecs/archetype.h"
+#include "plex/ecs/view_relations.h"
 
 #include <gtest/gtest.h>
 
 namespace plex::tests
 {
+using namespace ::plex::details;
+
+namespace
+{
+  template<size_t Tag>
+  struct TestType
+  {};
+} // namespace
+
+TEST(ViewRelations_Tests, GetComponentId_Single_AlwaysSame)
+{
+  EXPECT_EQ(GetComponentId<TestType<0>>(), GetComponentId<TestType<0>>());
+}
+
+TEST(ViewRelations_Tests, GetComponentId_Double_Different)
+{
+  EXPECT_NE(GetComponentId<TestType<0>>(), GetComponentId<TestType<1>>());
+}
+
+TEST(ViewRelations_Tests, GetViewId_Single_AlwaysSame)
+{
+  EXPECT_EQ(GetViewId<TestType<0>>(), GetViewId<TestType<0>>());
+}
+
+TEST(ViewRelations_Tests, GetViewId_Double_Different)
+{
+  EXPECT_NE(GetViewId<TestType<0>>(), GetViewId<TestType<1>>());
+}
+
+TEST(ViewRelations_Tests, GetArchetypeId_Single_AlwaysSame)
+{
+  EXPECT_EQ(GetArchetypeId<TestType<0>>(), GetArchetypeId<TestType<0>>());
+}
+
+TEST(ViewRelations_Tests, GetArchetypeId_Double_Different)
+{
+  EXPECT_NE(GetArchetypeId<TestType<0>>(), GetArchetypeId<TestType<1>>());
+}
+
+TEST(ViewRelations_Tests, GetComponentIds_Single_Same)
+{
+  auto list = GetComponentIds<TestType<0>>();
+
+  EXPECT_EQ(GetComponentId<TestType<0>>(), list[0]);
+}
+
+TEST(ViewRelations_Tests, GetComponentIds_Multiple_Same)
+{
+  auto list = GetComponentIds<TestType<0>, TestType<1>, TestType<2>>();
+
+  EXPECT_NE(std::ranges::find(list, GetComponentId<TestType<0>>()), list.end());
+  EXPECT_NE(std::ranges::find(list, GetComponentId<TestType<1>>()), list.end());
+  EXPECT_NE(std::ranges::find(list, GetComponentId<TestType<2>>()), list.end());
+  EXPECT_EQ(std::ranges::find(list, GetComponentId<TestType<3>>()), list.end());
+}
+
+TEST(ViewRelations_Tests, GetComponentIds_ObtainedTwiceSameOrder_Same)
+{
+  const auto& list1 = GetComponentIds<TestType<0>, TestType<1>, TestType<2>>();
+  const auto& list2 = GetComponentIds<TestType<0>, TestType<1>, TestType<2>>();
+
+  EXPECT_EQ(list1, list2);
+}
+
+TEST(ViewRelations_Tests, GetComponentIds_ObtainedTwiceDifferentOrder_Same)
+{
+  const auto& list1 = GetComponentIds<TestType<0>, TestType<1>, TestType<2>>();
+  const auto& list2 = GetComponentIds<TestType<2>, TestType<0>, TestType<1>>();
+
+  EXPECT_EQ(list1, list2);
+}
+
+TEST(ViewRelations_Tests, GetComponentIds_ObtainedTwiceDifferentValues_Different)
+{
+  const auto& list1 = GetComponentIds<TestType<0>, TestType<1>, TestType<2>>();
+  const auto& list2 = GetComponentIds<TestType<2>, TestType<5>, TestType<1>>();
+
+  EXPECT_NE(list1, list2);
+}
+
 TEST(ViewRelations_Tests, AssureArchetype_Single_UniqueId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_NE(relations.AssureArchetype<>(), relations.AssureArchetype<int>());
   EXPECT_NE(relations.AssureArchetype<double>(), relations.AssureArchetype<int>());
@@ -14,7 +94,7 @@ TEST(ViewRelations_Tests, AssureArchetype_Single_UniqueId)
 
 TEST(ViewRelations_Tests, AssureArchetype_SingleTwice_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ(relations.AssureArchetype<int>(), relations.AssureArchetype<int>());
   EXPECT_EQ(relations.AssureArchetype<float>(), relations.AssureArchetype<float>());
@@ -22,7 +102,7 @@ TEST(ViewRelations_Tests, AssureArchetype_SingleTwice_SameId)
 
 TEST(ViewRelations_Tests, AssureArchetype_MultipleSameOrder_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ((relations.AssureArchetype<int, double>()), (relations.AssureArchetype<int, double>()));
   EXPECT_EQ((relations.AssureArchetype<int, double, float>()), (relations.AssureArchetype<int, double, float>()));
@@ -30,7 +110,7 @@ TEST(ViewRelations_Tests, AssureArchetype_MultipleSameOrder_SameId)
 
 TEST(ViewRelations_Tests, AssureArchetype_MultipleDifferentOrder_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ((relations.AssureArchetype<int, double>()), (relations.AssureArchetype<double, int>()));
   EXPECT_EQ((relations.AssureArchetype<float, double, int>()), (relations.AssureArchetype<int, float, double>()));
@@ -38,7 +118,7 @@ TEST(ViewRelations_Tests, AssureArchetype_MultipleDifferentOrder_SameId)
 
 TEST(ViewRelations_Tests, AssureArchetype_MultipleDifferentValues_DifferentIds)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_NE((relations.AssureArchetype<int, double>()), (relations.AssureArchetype<int>()));
   EXPECT_NE((relations.AssureArchetype<int, float>()), (relations.AssureArchetype<double, int>()));
@@ -47,7 +127,7 @@ TEST(ViewRelations_Tests, AssureArchetype_MultipleDifferentValues_DifferentIds)
 
 TEST(ViewRelations_Tests, AssureView_Single_UniqueId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_NE(relations.AssureView<>(), relations.AssureView<int>());
   EXPECT_NE(relations.AssureView<double>(), relations.AssureView<int>());
@@ -55,7 +135,7 @@ TEST(ViewRelations_Tests, AssureView_Single_UniqueId)
 
 TEST(ViewRelations_Tests, AssureView_SingleTwice_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ(relations.AssureView<int>(), relations.AssureView<int>());
   EXPECT_EQ(relations.AssureView<float>(), relations.AssureView<float>());
@@ -63,7 +143,7 @@ TEST(ViewRelations_Tests, AssureView_SingleTwice_SameId)
 
 TEST(ViewRelations_Tests, AssureView_MultipleSameOrder_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ((relations.AssureView<int, double>()), (relations.AssureView<int, double>()));
   EXPECT_EQ((relations.AssureView<int, double, float>()), (relations.AssureView<int, double, float>()));
@@ -71,7 +151,7 @@ TEST(ViewRelations_Tests, AssureView_MultipleSameOrder_SameId)
 
 TEST(ViewRelations_Tests, AssureView_MultipleDifferentOrder_SameId)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_EQ((relations.AssureView<int, double>()), (relations.AssureView<double, int>()));
   EXPECT_EQ((relations.AssureView<float, double, int>()), (relations.AssureView<int, float, double>()));
@@ -79,7 +159,7 @@ TEST(ViewRelations_Tests, AssureView_MultipleDifferentOrder_SameId)
 
 TEST(ViewRelations_Tests, AssureView_MultipleDifferentValues_DifferentIds)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   EXPECT_NE((relations.AssureView<int, double>()), (relations.AssureView<int>()));
   EXPECT_NE((relations.AssureView<int, float>()), (relations.AssureView<double, int>()));
@@ -88,7 +168,7 @@ TEST(ViewRelations_Tests, AssureView_MultipleDifferentValues_DifferentIds)
 
 TEST(ViewRelations_Tests, ViewArchetypes_Multiple_CorrectSize)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   relations.AssureArchetype<int>();
   relations.AssureArchetype<float>();
@@ -117,7 +197,7 @@ TEST(ViewRelations_Tests, ViewArchetypes_Multiple_CorrectSize)
 
 TEST(ViewRelations_Tests, ViewArchetypes_AfterView_CorrectSize)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   relations.AssureArchetype<int>();
   relations.AssureArchetype<float>();
@@ -138,7 +218,7 @@ TEST(ViewRelations_Tests, ViewArchetypes_AfterView_CorrectSize)
 
 TEST(ViewRelations_Tests, ViewArchetypes_Multiple_CorrectArchetypes)
 {
-  ArchetypeViewMappings relations;
+  ViewRelations relations;
 
   Vector<ArchetypeId> archetypes;
   Vector<ArchetypeId> view_archetypes;
